unlock_tool/mediatek_brom_sample.py
# MediaTek BROM Exploit Sample - Python Implementation
# This demonstrates the basic structure of a MediaTek BootROM exploit

import usb.core
import usb.util
import struct
import time
from typing import Optional, Tuple

class MediaTekBROMExploit:
    """Sample MediaTek BootROM exploit implementation"""

    def __init__(self):
        self.device = None
        self.chip_info = {}

    def find_device(self) -> bool:
        """Find MediaTek device in BROM mode"""
        self.device = usb.core.find(idVendor=0x0e8d, idProduct=0x0003)
        if self.device is None:
            self.device = usb.core.find(idVendor=0x0e8d, idProduct=0x2000)

        if self.device is None:
            print("No MediaTek BROM device found")
            return False

        try:
            self.device.set_configuration()
            print(f"Found MediaTek device: {self.device.product}")
            return True
        except usb.core.USBError as e:
            print(f"USB configuration error: {e}")
            return False

    def send_handshake(self) -> bool:
        """Send initial handshake to BROM"""
        if not self.device:
            return False

        try:
            # Send handshake packet
            handshake = b'\xA0\x0A\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00'
            self.device.write(0x01, handshake)

            # Read response
            response = self.device.read(0x81, 64, timeout=1000)
            if len(response) >= 4:
                print(f"BROM handshake response: {response[:4].hex()}")
                return True
        except usb.core.USBError as e:
            print(f"Handshake failed: {e}")

        return False

    def read_chip_info(self) -> Optional[dict]:
        """Read chip information from BROM"""
        if not self.device:
            return None

        try:
            # Send read chip info command
            cmd = struct.pack('<I', 0xFE000000)  # Read chip info command
            self.device.write(0x01, cmd)

            # Read response
            response = self.device.read(0x81, 32, timeout=1000)
            if len(response) >= 8:
                hw_code = struct.unpack('<I', response[0:4])[0]
                sw_ver = struct.unpack('<I', response[4:8])[0]

                self.chip_info = {
                    'hw_code': f'0x{hw_code:08X}',
                    'sw_version': f'0x{sw_ver:08X}',
                    'chip_name': self._identify_chip(hw_code)
                }

                print(f"Chip Info: {self.chip_info}")
                return self.chip_info

        except usb.core.USBError as e:
            print(f"Failed to read chip info: {e}")

        return None

    def _identify_chip(self, hw_code: int) -> str:
        """Identify chip based on hardware code"""
        chip_map = {
            0x6580: "MT6580",
            0x6735: "MT6735",
            0x6752: "MT6752",
            0x6761: "MT6761",
            0x6765: "MT6765",
            0x6771: "MT6771",
            0x6785: "MT6785",
            0x6833: "MT6833",
            0x6853: "MT6853",
            0x6873: "MT6873",
            0x6885: "MT6885",
            0x6893: "MT6893",
            0x6983: "MT6983",
            0x6985: "MT6985"
        }
        return chip_map.get(hw_code, f"Unknown (0x{hw_code:04X})")

    def exploit_brom(self) -> bool:
        """Perform the actual BROM exploit"""
        if not self.device or not self.chip_info:
            return False

        try:
            # Send exploit payload to crash security watchdog
            exploit_payload = self._generate_exploit_payload()

            # Write payload to specific memory address
            self._write_memory(0x00020000, exploit_payload)

            # Trigger the exploit
            self._trigger_exploit()

            print("BROM exploit completed successfully")
            return True

        except Exception as e:
            print(f"BROM exploit failed: {e}")
            return False

    def _generate_exploit_payload(self) -> bytes:
        """Generate the exploit payload"""
        # This is a simplified example - real exploits are much more complex
        return b'\x00' * 1024  # Placeholder

    def _write_memory(self, address: int, data: bytes):
        """Write data to device memory"""
        # Implementation would send write commands to BROM
        pass

    def _trigger_exploit(self):
        """Trigger the actual exploit"""
        # Implementation would send the final trigger command
        pass

def main():
    """Main function demonstrating MediaTek BROM exploit"""
    print("MediaTek BROM Exploit Sample")
    print("=" * 40)

    exploit = MediaTekBROMExploit()

    if not exploit.find_device():
        return

    if not exploit.send_handshake():
        return

    chip_info = exploit.read_chip_info()
    if not chip_info:
        return

    print(f"Ready to exploit {chip_info['chip_name']}")

    # Note: Actual exploitation requires specific payloads for each chip
    # This is just a demonstration of the communication structure

if __name__ == "__main__":
    main()