unlock_tool/requirements.txt
PyQt6>=6.5.0
pyusb>=1.3.0
pyserial>=3.5
libusb1>=3.0.0
adb-shell>=0.4.4
cryptography>=41.0.0
pycryptodome>=3.23.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
colorama>=0.4.6
tqdm>=4.65.0
pyyaml>=6.0
jsonschema>=4.17.0
psutil>=5.9.0
platformdirs>=3.8.0
packaging>=23.1
capstone>=5.0.0
keystone-engine>=0.9.2
qrcode>=7.4.0
passlib>=1.7.4
Exscript>=1.6.0
frida>=16.0.0
lief>=0.14.0

